Make the display update like a calculator
Updated the event listener for button 1 with an if statement, so that
if the value in the array is 0, the array is cleared when the button
is pressed and replaced with the button figure, in this case 1. Then
on each additional press, the button adds another 1 as an element to
the array. Changed target to "screen".

Switched the display screen class to an id called "screen".