Create add button function
Added a new id to the add button. The button uses an if statement
to check if the value in numOne is empty or not, if it is then it will
call operate and add numOne and numTwo together and move the result into
numOne and emptying numTwo, whilst updating the display value.