Make it so answer can be used in calculations
Was having an issue with the answer from the previous sum not being
useable in the next calculation. Fixed it by adding an else if statement
that checked if numOne is not empty and if operator was empty, and if so
to update the operator and clear displayValue. Added id to equals button
in index.html.