Add backspace, fix divide by zero issue
Added backspace function, fairly simple just pops the last result of the
displayValue array and updates it.

Message wasn't displaying if divide by zero, changed operate function
so that it performs a check before it returns the divide result, if the
result is a string it won't try to round it up etc.